+++ src/types-internal/editor-modules.d.ts
@@ -37,6 +37,7 @@ import ConversionToolbar from '../components/modules/toolbar/conversion';
 import TooltipAPI from '../components/modules/api/tooltip';
 import ReadOnly from '../components/modules/readonly';
 import ReadOnlyAPI from '../components/modules/api/readonly';
+import I18nAPI from '../components/modules/api/i18n';
 
 export interface EditorModules {
   UI: UI;
@@ -78,4 +79,5 @@ export interface EditorModules {
   TooltipAPI: TooltipAPI;
   ReadOnly: ReadOnly;
   ReadOnlyAPI: ReadOnlyAPI;
+  I18nAPI: I18nAPI;
 }