+++ types/index.d.ts
@@ -4,7 +4,13 @@
  * ------------------------------------
  */
 
-import {EditorConfig} from './configs';
+import {
+  EditorConfig,
+  I18nDictionary,
+  Dictionary,
+  DictValue,
+  I18nConfig,
+} from './configs';
 import {
   Blocks,
   Caret,
@@ -18,6 +24,7 @@ import {
   Styles,
   Toolbar,
   Tooltip,
+  I18n,
 } from './api';
 import {OutputData} from './data-formats/output-data';
 
@@ -47,7 +54,17 @@ export {
   FilePasteEventDetail,
 } from './tools';
 export {BlockTune, BlockTuneConstructable} from './block-tunes';
-export {EditorConfig, SanitizerConfig, PasteConfig, LogLevels, ConversionConfig} from './configs';
+export {
+  EditorConfig,
+  SanitizerConfig,
+  PasteConfig,
+  LogLevels,
+  ConversionConfig,
+  I18nDictionary,
+  Dictionary,
+  DictValue,
+  I18nConfig,
+} from './configs';
 export {OutputData} from './data-formats/output-data';
 
 /**
@@ -67,6 +84,7 @@ export interface API {
   toolbar: Toolbar;
   inlineToolbar: InlineToolbar;
   tooltip: Tooltip;
+  i18n: I18n;
 }
 
 /**