+++ docs/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - `Improvements` - TSLint (deprecated) replaced with ESLint, old config changed to [CodeX ESLint Config](https://github.com/codex-team/eslint-config).
 - `Improvements` - Fix many code-style issues, add missed annotations.
-- `Improvements` - Adjusted GitHub action for ESLint.
+- `Improvements` - Adjusted GitHub Action for ESLint.
+- `New` *I18n API* — Ability to provide internalization for Editor.js core and tools. [#751](https://github.com/codex-team/editor.js/issues/751)
 
 ### 2.17